requirements.txt
pandas
pyarrow
pytest
pytest-cov
openpyxl
psycopg2-binary
pymysql
sqlalchemy
pre-commit
black
flake8
isort
mypy
